+++ lingui.config.cjs
@@ -1,0 +1,16 @@
+const { formatter } = require("@lingui/format-po");
+
+/** @type {import('@lingui/conf').LinguiConfig} */
+module.exports = {
+  locales: ["en", "de"],
+  sourceLocale: "en",
+  catalogs: [
+    {
+      path: "src/locales/{locale}/messages",
+      include: ["src"],
+      exclude: ["**/*.d.ts"],
+    },
+  ],
+  format: formatter({ lineNumbers: false, explicitIdAsDefault: true }),
+  orderBy: "messageId",
+};